Add Command class
