+++ src/test/java/command/InputTypesTest.java
@@ -1,0 +1,4 @@
+package command;
+
+public class InputTypesTest {
+}
